refactor(common/sst): moved DRRAOutput class to a separate header file and made Instruction::get method const
